feat: group Info action 과 Link action 이랑 분리
