+++ src/components/page/meetingDetail/Feed/FeedPanel.tsx
@@ -13,6 +13,7 @@ import { useRouter } from 'next/router';
 import { styled } from 'stitches.config';
 import EmptyView from './EmptyView';
 import FeedItem from './FeedItem';
+import GroupInfo from './FeedItem/GroupInfo';
 import MobileFeedListSkeleton from './Skeleton/MobileFeedListSkeleton';
 
 interface FeedPanelProps {
@@ -63,7 +64,7 @@ const FeedPanel = ({ isMember }: FeedPanelProps) => {
       <a>
         {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
         {/* @ts-ignore */}
-        <FeedItem {...post!} />
+        <FeedItem post={post} HeaderSection={<GroupInfo />} />
       </a>
     </Link>
   ));
